Fix cylinder cap normal calculation
Until now we returned the same vector facing in a fixed direction for
both cylinder caps. This was causing that the bottom cap for every
cylinder could never be lit.

We fixed this by returning an outward facing vector as surface normal
for every cap (meaning we return an inverted axis if we hit the bottom
cap).